basic subsonic 1.12.0 api support. add subsonic api bookmarks support. add save/get playqueue support to both subsonic and shiva style apis. add auto-save and restore playqueue support to web ui (via shiva style api)

+++ businesslayer/playlistbusinesslayer.php
@@ -18,10 +18,13 @@ use \OCA\Music\AppFramework\Core\Logger;
 
 use \OCA\Music\Db\PlaylistMapper;
 use \OCA\Music\Db\Playlist;
+use \OCA\Music\Db\SortBy;
 
 use \OCA\Music\Utility\Util;
 
 class PlaylistBusinessLayer extends BusinessLayer {
+        const PLAYQUEUE_PLAYLIST_NAME = '{playqueue}';
+
 	private $logger;
 	private $trackBusinessLayer;
 
@@ -62,12 +65,34 @@ class PlaylistBusinessLayer extends BusinessLayer {
 	}
 
 	public function create($name, $userId) {
+    // if playlist with given name already exists, don't re-create it (return record will include existing tracks)
+    $existingPlaylists = $this->findAllByName($name, $userId);
+    if (count($existingPlaylists) > 0)
+      return $existingPlaylists[0];
+
 		$playlist = new Playlist();
 		$playlist->setName(Util::truncate($name, 256)); // some DB setups can't truncate automatically to column max size
 		$playlist->setUserId($userId);
 
 		return $this->mapper->insert($playlist);
 	}
+
+        public function createWithTracks($userId, $name, $trackIds) {
+          // if playlist with given name already exists, don't re-create it (return record will include existing tracks)
+          $existingPlaylists = $this->findAllByName($name, $userId);
+          if (count($existingPlaylists) > 0)
+          {
+            $existingPlaylists[0]->setTrackIdsFromArray($trackIds);
+            return $this->mapper->update($existingPlaylists[0]);
+          }
+
+          $playlist = new Playlist();
+          $playlist->setName(Util::truncate($name, 256)); // some DB setups can't truncate automatically to column max size
+          $playlist->setUserId($userId);
+          $playlist->setTrackIdsFromArray($trackIds);
+
+          return $this->mapper->insert($playlist);
+        }
 
 	public function rename($name, $playlistId, $userId) {
 		$playlist = $this->find($playlistId, $userId);
@@ -124,4 +149,56 @@ class PlaylistBusinessLayer extends BusinessLayer {
 
 		return $playlistTracks;
 	}
+
+	/**
+	 * @param string $name
+	 * @param string $userId
+	 * @param bool $fuzzy
+	 * @param integer $limit
+	 * @param integer $offset
+	 * @return Playlist[]
+	 */
+        public function findAllByName($name, $userId, $fuzzy = false, $limit=null, $offset=null) {
+          return $this->mapper->findAllByName($name, $userId, $fuzzy, $limit, $offset);
+        }
+
+	/**
+	 * Finds all entities
+	 * @param string $userId the name of the user
+	 * @param integer $sortBy sort order of the result set
+	 * @param integer $limit
+	 * @param integer $offset
+	 * @return Entity[]
+	 */
+        public function findAll($userId, $sortBy=SortBy::None, $limit=null, $offset=null) {
+          $playlists = $this->mapper->findAll($userId, $sortBy, $limit, $offset);
+
+          // remove playqueue if it exists in results
+          foreach ($playlists as $key => $playlist) {
+            if ($playlist->getName() == self::PLAYQUEUE_PLAYLIST_NAME) {
+              unset($playlists[$key]);
+              break;
+            }
+          }
+
+          return $playlists;
+        }
+
+	/**
+	 * Finds all entities
+	 * @param string $userId the name of the user
+	 * @return Entity[] or null
+	 */
+        public function findPlayQueue($userId) {
+          return array_shift($this->findAllByName(self::PLAYQUEUE_PLAYLIST_NAME, $userId));
+        }
+
+	/**
+	 * Finds all entities
+	 * @param string $userId the name of the user
+	 * @return Entity[] or null
+	 */
+        public function createPlayQueueWithTracks($userId, $trackIds) {
+          return $this->createWithTracks($userId, self::PLAYQUEUE_PLAYLIST_NAME, $trackIds);
+        }
 }
